a2a_scraper/requirements.txt
pydantic-ai 
mcp 
pydantic-graph

arxiv
google-generativeai
python-dotenv
requests
rich
asyncclick
flask

a2a-sdk

langchain-google-genai 
langchain 
langgraph

google-genai 
httpx 
uvicorn 
click 
rich

crawl4ai

selenium